[iOS] Move Pull to Refresh, missed values and delegate out of Tab
