+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Tab.swift
@@ -46,24 +46,6 @@ protocol TabContentScript: TabContentScriptLoader {
   )
 }
 
-protocol TabDelegate {
-  func showRequestRewardsPanel(_ tab: Tab)
-  func stopMediaPlayback(_ tab: Tab)
-  func showWalletNotification(_ tab: Tab, origin: URLOrigin)
-  func updateURLBarWalletButton()
-  func isTabVisible(_ tab: Tab) -> Bool
-}
-
-struct RewardsTabChangeReportingState {
-  /// Set to true when the resulting page was restored from session state.
-  var wasRestored = false
-  /// Set to true when the resulting page navigation is not a reload or a
-  /// back/forward type.
-  var isNewNavigation = true
-  /// HTTP status code of the resulting page.
-  var httpStatusCode = -1
-}
-
 enum TabSecureContentState: String {
   case unknown = "Unknown"
   case localhost = "Localhost"
@@ -102,12 +84,11 @@ class Tab: NSObject {
 
   private(set) var type: TabType = .regular
 
-  var redirectChain = [URL]()
-  var responses = [URL: URLResponse]()
-
   var isPrivate: Bool {
     return type.isPrivate
   }
+
+  var isVisible: Bool = false
 
   private var webViewObservations: [AnyCancellable] = []
 
@@ -223,7 +204,6 @@ class Tab: NSObject {
   var sampledPageTopColor: UIColor? {
     webView?.sampledPageTopColor
   }
-  var tabDelegate: TabDelegate?
   var favicon: Favicon
   var lastExecutedTime: Timestamp?
   fileprivate var lastRequest: URLRequest?
@@ -342,10 +322,6 @@ class Tab: NSObject {
     self.uiHandler = TabWKUIHandler(tab: self)
   }
 
-  private lazy var refreshControl = UIRefreshControl().then {
-    $0.addTarget(self, action: #selector(reload), for: .valueChanged)
-  }
-
   var isWebViewCreated: Bool {
     webView != nil
   }
@@ -417,7 +393,7 @@ class Tab: NSObject {
       url = newURL
       notifyObservers = true
     } else {
-      if tabDelegate?.isTabVisible(self) == true, url?.displayURL != nil,
+      if isVisible, url?.displayURL != nil,
         url?.displayURL?.scheme == "about", !loading
       {
         if let newURL {
@@ -426,8 +402,6 @@ class Tab: NSObject {
         }
       }
     }
-
-    updatePullToRefreshVisibility()
 
     if notifyObservers {
       observers.forEach {
@@ -737,53 +711,8 @@ class Tab: NSObject {
     return webView?.title
   }
 
-  var displayTitle: String {
-    if let displayTabTitle = fetchDisplayTitle(using: url, title: title) {
-      return displayTabTitle
-    }
-
-    // When picking a display title. Tabs with sessionData are pending a restore so show their old title.
-    // To prevent flickering of the display title. If a tab is restoring make sure to use its lastTitle.
-    if let url = self.url, InternalURL(url)?.isAboutHomeURL ?? false, !isRestoring {
-      return Strings.Hotkey.newTabTitle
-    }
-
-    if let url = self.url, !InternalURL.isValid(url: url),
-      let shownUrl = url.displayURL?.absoluteString, webView != nil
-    {
-      return shownUrl
-    }
-
-    guard let lastTitle = lastTitle, !lastTitle.isEmpty else {
-      // FF uses url?.displayURL?.absoluteString ??  ""
-      if let title = url?.absoluteString {
-        return title
-      } else if let tab = SessionTab.from(tabId: id) {
-        if tab.title.isEmpty {
-          return Strings.Hotkey.newTabTitle
-        }
-        return tab.title
-      }
-
-      return ""
-    }
-
-    return lastTitle
-  }
-
   var currentInitialURL: URL? {
     return self.webView?.backForwardList.currentItem?.initialURL
-  }
-
-  var displayFavicon: Favicon? {
-    if let url = url, InternalURL(url)?.isAboutHomeURL == true {
-      return Favicon(
-        image: UIImage(sharedNamed: "brave.logo"),
-        isMonogramImage: false,
-        backgroundColor: .clear
-      )
-    }
-    return favicon
   }
 
   var canGoBack: Bool {
@@ -792,42 +721,6 @@ class Tab: NSObject {
 
   var canGoForward: Bool {
     return webView?.canGoForward ?? false
-  }
-
-  /// This property is for fetching the actual URL for the Tab
-  /// In private browsing the URL is in memory but this is not the case for normal mode
-  /// For Normal  Mode Tab information is fetched using Tab ID from
-  var fetchedURL: URL? {
-    if isPrivate {
-      if let url = url, url.isWebPage() {
-        return url
-      }
-    } else {
-      if let tabUrl = url, tabUrl.isWebPage() {
-        return tabUrl
-      } else if let fetchedTab = SessionTab.from(tabId: id), fetchedTab.url?.isWebPage() == true {
-        return url
-      }
-    }
-
-    return nil
-  }
-
-  func fetchDisplayTitle(using url: URL?, title: String?) -> String? {
-    if let tabTitle = title, !tabTitle.isEmpty {
-      var displayTitle = tabTitle
-
-      // Checking host is "localhost" || host == "127.0.0.1"
-      // or hostless URL (iOS forwards hostless URLs (e.g., http://:6571) to localhost.)
-      // DisplayURL will retrieve original URL even it is redirected to Error Page
-      if let isLocal = url?.displayURL?.isLocal, isLocal {
-        displayTitle = ""
-      }
-
-      return displayTitle
-    }
-
-    return nil
   }
 
   func goBack() {
@@ -878,14 +771,6 @@ class Tab: NSObject {
     // loaded wrong user agent.
     webView?.customUserAgent = nil
 
-    defer {
-      if let refreshControl = webView?.scrollView.refreshControl,
-        refreshControl.isRefreshing
-      {
-        refreshControl.endRefreshing()
-      }
-    }
-
     // If the current page is an error page, and the reload button is tapped, load the original URL
     if let url = webView?.url, let internalUrl = InternalURL(url),
       let page = internalUrl.originalURLFromErrorPage
@@ -921,34 +806,6 @@ class Tab: NSObject {
     webView.customUserAgent = desktopMode ? UserAgent.desktop : UserAgent.mobile
   }
 
-  func hideContent(_ animated: Bool = false) {
-    webView?.isUserInteractionEnabled = false
-    if animated {
-      UIView.animate(
-        withDuration: 0.25,
-        animations: { () -> Void in
-          self.webView?.alpha = 0.0
-        }
-      )
-    } else {
-      webView?.alpha = 0.0
-    }
-  }
-
-  func showContent(_ animated: Bool = false) {
-    webView?.isUserInteractionEnabled = true
-    if animated {
-      UIView.animate(
-        withDuration: 0.25,
-        animations: { () -> Void in
-          self.webView?.alpha = 1.0
-        }
-      )
-    } else {
-      webView?.alpha = 1.0
-    }
-  }
-
   /// Switches user agent Desktop -> Mobile or Mobile -> Desktop.
   func switchUserAgent() {
     if let urlString = webView?.url?.baseDomain {
@@ -964,18 +821,8 @@ class Tab: NSObject {
     reload()
   }
 
-  func updatePullToRefreshVisibility() {
-    guard let url = webView?.url, let webView = webView else { return }
-    webView.scrollView.refreshControl =
-      url.isLocalUtility || !Preferences.General.enablePullToRefresh.value ? nil : refreshControl
-  }
-
   func isDescendentOf(_ ancestor: Tab) -> Bool {
     return sequence(first: parent) { $0?.parent }.contains { $0 == ancestor }
-  }
-
-  func stopMediaPlayback() {
-    tabDelegate?.stopMediaPlayback(self)
   }
 
   private func resolvedPolicyDecision(
@@ -1138,79 +985,6 @@ extension TabWebView: PreferencesObserver {
   func preferencesDidChange(for key: String) {
     updateBackgroundColor()
   }
-}
-
-// MARK: - Brave Search
-
-extension Tab {
-  /// Call the api on the Brave Search website and passes the fallback results to it.
-  /// Important: This method is also called when there is no fallback results
-  /// or when the fallback call should not happen at all.
-  /// The website expects the iOS device to always call this method(blocks on it).
-  func injectResults() {
-    DispatchQueue.main.async {
-      // If the backup search results happen before the Brave Search loads
-      // The method we pass data to is undefined.
-      // For such case we do not call that method or remove the search backup manager.
-
-      self.webView?.evaluateJavaScript("window.onFetchedBackupResults === undefined") {
-        result,
-        error in
-
-        if let error = error {
-          Logger.module.error(
-            "onFetchedBackupResults existence check error: \(error.localizedDescription, privacy: .public)"
-          )
-        }
-
-        guard let methodUndefined = result as? Bool else {
-          Logger.module.error(
-            "onFetchedBackupResults existence check, failed to unwrap bool result value"
-          )
-          return
-        }
-
-        if methodUndefined {
-          Logger.module.info("Search Backup results are ready but the page has not been loaded yet")
-          return
-        }
-
-        var queryResult = "null"
-
-        if let url = self.url,
-          BraveSearchManager.isValidURL(url),
-          let result = self.braveSearchManager?.fallbackQueryResult
-        {
-          queryResult = result
-        }
-
-        self.evaluateSafeJavaScript(
-          functionName: "window.onFetchedBackupResults",
-          args: [queryResult],
-          contentWorld: BraveSearchScriptHandler.scriptSandbox,
-          escapeArgs: false
-        )
-
-        // Cleanup
-        self.braveSearchManager = nil
-      }
-    }
-  }
-}
-
-// MARK: - Brave SKU
-extension Tab {
-  func injectLocalStorageItem(key: String, value: String) {
-    self.evaluateSafeJavaScript(
-      functionName: "localStorage.setItem",
-      args: [key, value],
-      contentWorld: BraveSkusScriptHandler.scriptSandbox
-    )
-  }
-}
-
-// MARK: Script Injection
-extension Tab {
 }
 
 // Find In Page interaction
